move video file extension to along with the livePhoto util
